remove SonarLint and Roslyn profile exporters
