--- sonar-dotnet-shared-library/src/main/java/org/sonarsource/dotnet/shared/plugins/SonarLintProfileExporter.java
@@ -1,31 +1,0 @@
-/*
- * SonarSource :: .NET :: Shared library
- * Copyright (C) 2014-2024 SonarSource SA
- * mailto:info AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-package org.sonarsource.dotnet.shared.plugins;
-
-import org.sonar.api.rules.RuleFinder;
-
-public class SonarLintProfileExporter extends AbstractSonarLintProfileExporter {
-  private static final String PROFILE_NAME = "SonarLint for Visual Studio Rule Set";
-
-  public SonarLintProfileExporter(RuleFinder ruleFinder, DotNetPluginMetadata pluginMetadata) {
-    super("sonarlint-vs-" + pluginMetadata.languageKey(), PROFILE_NAME, pluginMetadata.languageKey(),
-      pluginMetadata.sonarAnalyzerName(), pluginMetadata.repositoryKey(), ruleFinder);
-  }
-}
