+++ sonar-vbnet-plugin/src/main/java/org/sonar/plugins/vbnet/VbNetPlugin.java
@@ -34,8 +34,6 @@ import org.sonarsource.dotnet.shared.plugins.PropertiesSensor;
 import org.sonarsource.dotnet.shared.plugins.ProtobufDataImporter;
 import org.sonarsource.dotnet.shared.plugins.ReportPathCollector;
 import org.sonarsource.dotnet.shared.plugins.RoslynDataImporter;
-import org.sonarsource.dotnet.shared.plugins.RoslynProfileExporter;
-import org.sonarsource.dotnet.shared.plugins.SonarLintProfileExporter;
 import org.sonarsource.dotnet.shared.plugins.UnitTestResultsProvider;
 import org.sonarsource.dotnet.shared.plugins.WrongEncodingFileFilter;
 
@@ -81,15 +79,12 @@ public class VbNetPlugin implements Plugin {
       WrongEncodingFileFilter.class,
       // importers / exporters
       ProtobufDataImporter.class,
-      RoslynDataImporter.class,
-      RoslynProfileExporter.class,
-      SonarLintProfileExporter.class);
+      RoslynDataImporter.class);
 
     context.addExtensions(new VbNetPropertyDefinitions().create());
     context.addExtension(new VbNetSonarWayProfile());
     context.addExtensions(new CodeCoverageProvider(METADATA).extensions());
     context.addExtensions(new UnitTestResultsProvider(METADATA).extensions());
-    context.addExtensions(RoslynProfileExporter.sonarLintRepositoryProperties(METADATA));
   }
 
   private static class VbNetPluginMetadata implements DotNetPluginMetadata {
